Split build graph JSON validation tests
