+++ tests/integration/cli_build/build_graph_json/validation.rs
@@ -1,5 +1,9 @@
 use std::process::Command;
 
+#[path = "validation/dependencies.rs"]
+mod dependencies;
+#[path = "validation/deterministic_body.rs"]
+mod deterministic_body;
 #[path = "validation/unsupported_targets.rs"]
 mod unsupported_targets;
 
@@ -150,86 +154,6 @@ build = (b: Builder) Result<BuildConfig, BuildError> {
 }
 "#,
         "field `exports` in `Library` build target must contain at least one source",
-    );
-}
-
-#[test]
-fn emit_json_build_graph_rejects_unknown_target_dependencies() {
-    assert_emit_json_build_graph_error_contains(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        dependencies: ["core"],
-    })
-    .Ok(b.config())
-}
-"#,
-        "build target `app` depends on unknown target `core`",
-    );
-}
-
-#[test]
-fn emit_json_build_graph_rejects_self_target_dependencies() {
-    assert_emit_json_build_graph_error_contains(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        dependencies: ["app"],
-    })
-    .Ok(b.config())
-}
-"#,
-        "build target `app` cannot depend on itself",
-    );
-}
-
-#[test]
-fn emit_json_build_graph_rejects_cyclic_target_dependencies() {
-    assert_emit_json_build_graph_error_contains(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    b.add(Executable {
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        dependencies: ["tool"],
-    })
-    b.add(Executable {
-        name: "tool",
-        main: "tool.zen",
-        out_dir: "build/tool/",
-        dependencies: ["app"],
-    })
-    .Ok(b.config())
-}
-"#,
-        "build target dependency cycle includes `app`",
-    );
-}
-
-#[test]
-fn emit_json_build_graph_rejects_dynamic_target_adds() {
-    assert_emit_json_build_graph_error_contains(
-        r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {
-    enabled = true
-    enabled ?
-        | true {
-            b.add(Executable { name: "app", main: "app.zen", out_dir: "build/app/" })
-        }
-        | false {
-            b.add(Executable { name: "tool", main: "tool.zen", out_dir: "build/tool/" })
-        }
-    .Ok(b.config())
-}
-"#,
-        "build targets must be added in the deterministic build graph body",
     );
 }
